段错误

+++ database/db/test_cases.sql
@@ -1,7 +1,7 @@
 -- 测试用例表
 CREATE TABLE test_cases (
-    id INT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
-    problem_id INT UNSIGNED NOT NULL,           -- 关联的题目ID
+    id BIGINT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
+    problem_id BIGINT UNSIGNED NOT NULL,           -- 关联的题目ID
     stdin TEXT NOT NULL,                        -- 测试用例输入
     expected_output TEXT NOT NULL,              -- 期望输出
     sequence INT UNSIGNED NOT NULL,
